Toggle visibility of special row in EditExtended fragment as needed
This took a lot of trial and error, because the onCreateView events and
such are not fired when changing the tabs (they are executed before
because Android always initializes the tabs that could be swiped to
before).
But `setUserVisibleHint` is triggered, and from there `refreshText` is
called to hide the 'special' row if it is not needed.